// apps/api/src/proto/v1/credential/credential.ts
export interface CredentialRegisterBeginRequest {
}

export interface CredentialRegisterBeginResponse {
  optionsJson: Uint8Array;
  sessionId: string;
  challenge: Uint8Array;
}

export interface CredentialRegisterFinishRequest {
  sessionId: string;
  encMk: Uint8Array;
  mkIv: Uint8Array;
  response: Uint8Array;
  masterKeys: CredentialRegisterFinishRequest_WrappedMasterKey[];
}

export interface CredentialRegisterFinishRequest_WrappedMasterKey {
  cipher: Uint8Array;
  iv: Uint8Array;
  recoveryTag: Uint8Array;
}

export interface CredentialRegisterFinishResponse {
  success: boolean;
  credentialId: string;
  status: string;
  statusCode: string;
}

export interface CredentialLoginBeginRequest {
  email: string;
}

export interface CredentialLoginBeginResponse {
  challenge: string;
  rpId: string;
  optionsJson: Uint8Array;
  sessionId: string;
  pid: Uint8Array;
}

export interface CredentialLoginFinishRequest {
  sessionId: string;
  response: Uint8Array;
}

export interface CredentialLoginFinishResponse {
  success: boolean;
  message: string;
  encMk: Uint8Array;
  prfSalt: Uint8Array;
  recoveryTags: Uint8Array;
}

export interface OpaqueRegisterBeginRequest {
  clientRegistrationState: Uint8Array;
  registrationRequest: Uint8Array;
}

export interface OpaqueRegisterBeginResponse {
  registrationResponse: string;
  serverSetup: string;
}

export interface OpaqueRegisterFinishRequest {
  registrationRecord: Uint8Array;
}

export interface OpaqueRegisterFinishResponse {
  status: string;
  statusCode: number;
  error?: string | undefined;
}

export interface OpaqueBatchRegisterBeginRequest {
  requests: OpaqueRegisterBeginRequest[];
}

export interface OpaqueBatchRegisterBeginRespose {
  responses: OpaqueRegisterBeginResponse[];
}

export interface OpaqueBatchRegisterFinishRequest {
  requests: OpaqueRegisterFinishRequest[];
}

export interface OpaqueBatchRegisterFinishResponse {
  responses: OpaqueRegisterFinishResponse[];
}

export interface OpaqueLoginBeginRequest {
  clientLoginState: Uint8Array;
  startLoginRequest: Uint8Array;
}

export interface OpaqueLoginBeginResponse {
  loginResponse: string;
  serverLoginState: string;
}

export interface OpaqueLoginFinishRequest {
  finishLoginRequest: Uint8Array;
  sessionKey: Uint8Array;
}

export interface OpaqueLoginFinishResponse {
  status: string;
  statusCode: number;
  error?: string | undefined;
  salt?: Uint8Array | undefined;
}

export interface CredentialStoreDeviceParams {
  deviceName: string;
  description?: string | undefined;
  publicKey: string;
  enabled: boolean;
}

export interface CredentialRetrieveDeviceParams {
  deviceName: string;
}

export interface CredentialListDevicesParams {
  active: boolean;
}

export interface CredentialDevice {
  name: string;
  publicKey: string;
  description?: string | undefined;
  active: boolean;
  enabled: boolean;
}

export interface CredentialDeviceResponse {
  devices: CredentialDevice[];
}

export interface CredentialStoreKeysParams {
  encodedBlob: Uint8Array;
  userKey?: CredentialStoreKeysParams_UserKey | undefined;
  sessionKey?: CredentialStoreKeysParams_SessionKey | undefined;
  deviceKey?: CredentialStoreKeysParams_DeviceKey | undefined;
  keyType: string;
  publicKey: Uint8Array;
  credentialId?: string | undefined;
  salt: string;
}

export interface CredentialStoreKeysParams_UserKey {
  masterKeyCipher: string;
  salt: string;
  keyCipher?: string | undefined;
  publicKey?: string | undefined;
}

export interface CredentialStoreKeysParams_DeviceKey {
  masterKeyCipher: string;
  salt: string;
  keyCipher?: string | undefined;
  publicKey?: string | undefined;
}

export interface CredentialStoreKeysParams_SessionKey {
  codeCipher: Uint8Array;
  salt: Uint8Array;
}

export interface CredentialRetrieveKeysParams {
  type: string;
}

export interface CredentialResponse {
  status: string;
  statusCode: number;
  userKey?: CredentialResponse_UserKey | undefined;
  sessionKey?: CredentialResponse_SessionKey | undefined;
  list?: CredentialDeviceResponse | undefined;
}

export interface CredentialResponse_UserKey {
}

export interface CredentialResponse_SessionKey {
}

export const V1_CREDENTIAL_PACKAGE_NAME = "v1.credential";

/** Booking Service */

export interface CredentialServiceClient {
  /** Store a Credential */

  storeDevice(request: CredentialStoreDeviceParams, metadata: Metadata, ...rest: any): Observable<CredentialResponse>;

  /** Retrieve a Credential */

  retrieveDevice(
    request: CredentialRetrieveDeviceParams,
    metadata: Metadata,
    ...rest: any
  ): Observable<CredentialResponse>;

  /** List Credentials */

  listDevices(request: CredentialListDevicesParams, metadata: Metadata, ...rest: any): Observable<CredentialResponse>;

  storeKeys(request: CredentialStoreKeysParams, metadata: Metadata, ...rest: any): Observable<CredentialResponse>;

  saveKey(request: CredentialStoreKeysParams, metadata: Metadata, ...rest: any): Observable<CredentialResponse>;

  retrieveKeys(request: CredentialRetrieveKeysParams, metadata: Metadata, ...rest: any): Observable<CredentialResponse>;

  registerBegin(request: Empty, metadata: Metadata, ...rest: any): Observable<CredentialRegisterBeginResponse>;

  registerFinish(
    request: CredentialRegisterFinishRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<CredentialRegisterFinishResponse>;

  loginBegin(
    request: CredentialLoginBeginRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<CredentialLoginBeginResponse>;

  loginFinish(
    request: CredentialLoginFinishRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<CredentialLoginFinishResponse>;

  opaqueRegisterBegin(
    request: OpaqueRegisterBeginRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<OpaqueRegisterBeginResponse>;

  opaqueRegisterFinish(
    request: OpaqueRegisterFinishRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<OpaqueRegisterFinishResponse>;

  opaqueLoginBegin(
    request: OpaqueLoginBeginRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<OpaqueLoginBeginResponse>;

  opaqueLoginFinish(
    request: OpaqueLoginFinishRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<OpaqueLoginFinishResponse>;

  opaqueBatchRegisterBegin(
    request: OpaqueBatchRegisterBeginRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<OpaqueBatchRegisterBeginRespose>;

  opaqueBatchRegisterFinish(
    request: OpaqueBatchRegisterFinishRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<OpaqueBatchRegisterFinishResponse>;
}
// ...
export const CREDENTIAL_SERVICE_NAME = "CredentialService";